site.categorymembers: Support cmstartsortkeyprefix API parameter
site.py:
The old `startsort` is passed to API as `cmstartsortkey` which is
deprecated and is not compatible with `cmstartsortkeyprefix`. Also
working with `startsort` is difficult as it should be a "binary string"
usually obtained via API or directly from database. (See T74101)

Add two new parameters, `startprefix` and `endprefix`, to be passed to API
as `cmstartsortkeyprefix` and `cmendsortkeyprefix`.

Fix the docstring for `startsort` and `endsort`. They are in binary string
format and cannot be compared with a title directly. Also mention that they
are deprecated in MW 1.24.

page.py:
Add the new parameters to `Category.articles`.

Bug: T170265
Change-Id: I08f4a7a4bb1bbe5c10fe678cc7a951a273f61d07