[IMPR] Enhance argument management in harvest_template.py
Now each param-property pair holds its OptionHandler where
options specific for this claim are stored. Additionally,
it's possible to apply an option to the bot which applies
the option to all claims. Each claim can override this option
if the input is correct (see the updated instructions).

The only setting is at the moment -islink, which makes the
feature added in I71f6857ca optional.

Bug: T64014
Bug: T72702
Change-Id: Ibc4a0f845fe88349d49b787b5f16b20eb650ef8e