site_tests.py: Handle random errors caused by page content change
If a page's content is changed during the tests, the set of its templates or
wiki-links may also change. This is a race condition which causes an
AssertionError and random failures as described in T92856, T153501 and
T169649. To remedy the issue apply the following changes.

tests.__init__:

    Sort the imports.

    Define `force_cache_update` as a context manager which can temporarily
    disable TestRequest's cache when required.

site_tests.py:

    Sort the imports.

    On AssertionErrors that might have been caused by race conditions, assert
    that the cause has really been a change in links/templates of the page
    during the test. Also update the cache for the next run.
    Use `force_cache_update` to bypass the existing cache.

    Remove the unittest_print and related comments that were previously added
    as part of T92856.

Bug: T92856
Bug: T153501
Bug: T169649
Change-Id: I3b243929a46fafc000b8f175cfa99d32b697941d