site_tests.py: Avoid race conditions in test_pagetemplates and test_pagelinks
Instead of comparing the results of two generator calls that might not match,
either because of random functions used in the page or simultaneous user
edits, just make sure that the correct parameters are assigned to the
underlying Request.

In addition, for namespace-restricted generators confirm that the results are
all from the expected namespace.

Bug: T92856
Bug: T153501
Bug: T169649
Change-Id: I775729eb7153168140dbc832e0ae15790dd66779