Fix up PYWIKIBOT2_DIR
get_base_dir.exists(directory) check whether a 'user-config.py' file
exists. However when the file has not been created yet, it returns
false.
On a new setup, setting PYWIKIBOT2_DIR is thus always skipped and
get_base_dir() fall back to the user homedirectory, which is not quite
what is intended.

Make exists() check for the directory instead of a file that has yet to
be created.

Change-Id: I8a2f13a1c437c0560ab33afba3cdf4f5404be73b