[compat] drop compat module
core branch is now 10 years old and decommission of compat was one year ago.
The compat module was created in 2012 and older catlib and userlib where
merged into that folder in 2014.

The framework is not using this compatibility modul anymore and bot
operators get a deprecation waring for their private scripts when the module
is used.

I guess it is time to drop this module because all scripts should use the
core framework equivalent parts. It might be more difficult to keep this
backward compatibility alive and btw its usage is unsure because the behavior
is never testet.

- drop compat library module
- remove references from docs
- remove folder it from pwb sys path
- update README-conversion.txt
- update compat2core.py maintenance script
- prepare history file for the next pypi release

Change-Id: I32cd9ca2d1ef32a25f741fe46160317cb649c4ef