diff_checker.py: Add a check for double quoted strings
The new check will fail if there is a double quoted string in the
newly added/ modified lines that does not contain any single quotes.
A single quoted string should be used in this situation.

Now that we are tokenizing the modules, we can rewrite the u-prefix
check using the new logic that will not have any false positive.

Use `line.path` instead of `line.target_file` in check check. The former
the actual file path. `target_file` starts with an unnecessary 'b/'
prefix.

Change-Id: I84afcc0817a48514d1a99e8cae2091847931aff3
Signed-off-by: dalba <dalba.wiki@gmail.com>