diff_checker.py: Fix the logic of calculating error state
Also fix an encoding issue in Python 3 where `universal_newlines` parameter
in `subprocess.check_output` forces the result to be in text mode, but it also
needs to know the `encoding` of the result to correctly decode utf-8 diffs,
but `check_output` in Python 2 does not have the `encoding` parameter.
Therefore we have to decode the file and normalize newlines ourselves in
Python 2 and 3.

Change-Id: I7ab6d63c1d213087e23481679c7ef2a8665835d9