tools/__init__.py: Make suppress_warnings also work as a decorator
`suppress_warnings` already worked as a context manager, but it could not be
configured to suppress specific warnings. The new `message`, `category`, and
`filename` parameters will allow that.

Make the exising usages of `suppress_warnings` more specific. Use it as a
decorator in site_tests.py to suppress an expected DeprecationWarning in
site.search.

Change-Id: Ic606d75f359c4d9c6bc74927da8e6cb0fdf67ab7