protect.py: Add expiry parameter
If the paramater is unset, use None, meaning indefinite and protection
will not expire automatically. If '-expiry' is set but the value is
unspecified, the script will prompt with pywikibot.input(). The value,
either specified directly or in by the prompt, will be passes to
page.protect(), which will then pass it on to site.protect().

Bug: T183796
Change-Id: I486b5b140c46460048d55d14e1be3309dc7e1744