TerminalHandler.emit: Do not omit the last line even if it seems to be warn()
In T184368 we have a case that useful warning info are removed from the output.
This patch removes the lines responsible for removing the last line of a
warning message that appears to be a warn call.

In Pywikibot we usually fix warning stacklevel inside the warning
decorators (e.g. in issue_deprecation_warning). Even if we did not, we can
fix the code. Therefore, it seems unnecessary to remove the last line,
especially with the rist of removing useful information.

Bug: T184368
Change-Id: Iaf4b1dfb8939ce76b670e44f882e45504279a60a